Refactor MedicalAnalysisService to reduce cyclomatic complexity.
- Fixed compilation error in `FilePickerService` due to incorrect API usage.
- Moved `ConfidenceThreshold` and `SafeAnalysisResponse` to `lib/features/medical_assistant/domain/entities/analysis_response.dart`.
- Decomposed `MedicalAnalysisService` logic into `LabAnalysisStrategy`, `VitalAnalysisStrategy`, and `SymptomAnalysisStrategy`.
- Updated `MedicalAnalysisService` to delegate to these new strategies while maintaining backward compatibility.
- Improved `LoincCode.fromJson` in `packages/medical_standards` to handle missing optional fields in the JSON data.
- Added comprehensive unit tests for `MedicalAnalysisService`.
