+++ windows/flutter/generated_plugins.cmake
@@ -10,7 +10,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage_windows
   isar_flutter_libs
   local_auth_windows
-  objectbox_flutter_libs
   permission_handler_windows
 )
